Fix VrlValue int→i64, rust-compat UUIDv7, timestamp precision, conformance +2
Major changes:
- Change VrlValue integer variant from int (32-bit) to i64 (64-bit) across
  the entire VRL codebase, fixing integer overflow/truncation for large values.
  Parser now uses .i64() instead of .int() for integer literals.
- Add rust_vrl_compat build-time flag (src/vrl/compat.v) that enables
  emulation of Rust VRL's UUIDv7 quirk: chrono's timestamp_nanos_opt()→u32
  truncation that shifts the 48-bit millisecond timestamp.
- Fix timestamp formatting to use sub-millisecond precision matching Rust
  chrono's SecondsFormat::AutoSi (0, 3, or 6 fractional digits as needed).
- Add special UUID v7 test handler in conformance tests that validates
  structure, masks random bits, verifies non-zero randomness, and adjusts
  timestamp prefix comparison based on rust_vrl_compat flag.
- Fix to_unix_timestamp to always return i64 instead of demoting to f64
  for values exceeding i32 range.

Conformance: 207 pass, 14 fail (was 205/16). Fixed tests:
  - uuid_v7_valid_with_timestamp (rust-compat UUIDv7 timestamp)
  - arithmetic/multiplication/overflow (i64 integer literals)

+++ src/vrl/stdlib_enumerate.v
@@ -18,15 +18,15 @@ fn fn_tally(args []VrlValue) !VrlValue {
 		if existing := result.get(key) {
 			e := existing
 			match e {
-				int {
+				i64 {
 					result.set(key, VrlValue(e + 1))
 				}
 				else {
-					result.set(key, VrlValue(1))
+					result.set(key, VrlValue(i64(1)))
 				}
 			}
 		} else {
-			result.set(key, VrlValue(1))
+			result.set(key, VrlValue(i64(1)))
 		}
 	}
 	return VrlValue(result)
@@ -49,7 +49,7 @@ fn fn_tally_value(args []VrlValue) !VrlValue {
 			count++
 		}
 	}
-	return VrlValue(count)
+	return VrlValue(i64(count))
 }
 
 // match_array(value, pattern)
